Add spack-lightweight package manager
This commit adds a spack-lightweight package manager to encapsulate most
of the spack package manager logic, but to allow for deferred
installation.
